requirements.txt
webdriver-manager
flask
selenium
mysql-connector-python